Rename FirstInputTimings file name
