+++ packages/rum-core/src/domain/view/viewMetrics/trackInitialViewMetrics.ts
@@ -3,8 +3,8 @@ import type { RumConfiguration } from '../../configuration'
 import type { LifeCycle } from '../../lifeCycle'
 import type { WebVitalTelemetryDebug } from '../startWebVitalTelemetryDebug'
 import { trackFirstContentfulPaint } from './trackFirstContentfulPaint'
-import type { FirstInput } from './trackFirstInputTimings'
-import { trackFirstInput } from './trackFirstInputTimings'
+import type { FirstInput } from './trackFirstInput'
+import { trackFirstInput } from './trackFirstInput'
 import type { NavigationTimings } from './trackNavigationTimings'
 import { trackNavigationTimings } from './trackNavigationTimings'
 import type { LargestContentfulPaint } from './trackLargestContentfulPaint'
